[API] Enable contract tests for Locale
